# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Formora API
  version: 0.7.0
  description: HTTP API for the Formora platform.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/workspaces:
    get:
      operationId: listWorkspaces
      summary: List workspaces available to the authenticated user
      tags: [Workspaces]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The user's workspace memberships.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceListResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
    post:
      operationId: createWorkspace
      summary: Create a workspace with the authenticated user as owner
      tags: [Workspaces]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: The workspace and owner membership were created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
  /api/v1/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      summary: Retrieve an authorized workspace
      tags: [Workspaces]
      security:
        - cookieAuth: []
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-fA-F0-9]{24}$"
      responses:
        "200":
          description: The requested workspace and the user's role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          description: The user is not a workspace member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: The workspace was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/forms/{formId}/submissions:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: listFormSubmissions
      summary: List authorized form responses with publication and date filters
      tags: [Forms]
      security:
        - cookieAuth: []
      parameters:
        - name: publicationVersion
          in: query
          schema:
            type: integer
            minimum: 1
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of submission summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedSubmissionListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/public/forms/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicForm
      summary: Retrieve a published form or its password challenge
      tags: [Public Forms]
      security: []
      responses:
        "200":
          description: The published form. Definition is null until a protected form is unlocked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicFormResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/public/forms/{slug}/access:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: unlockPublicForm
      summary: Exchange a form password for a short-lived access token
      tags: [Public Forms]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
                  maxLength: 128
      responses:
        "200":
          description: Password accepted and access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicFormAccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/public/forms/{slug}/submissions:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      - name: X-Form-Access-Token
        in: header
        required: false
        description: Required for password-protected forms after unlocking.
        schema:
          type: string
    post:
      operationId: submitPublicForm
      summary: Validate and store answers against the published snapshot
      tags: [Public Forms]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicSubmissionRequest"
      responses:
        "201":
          description: The response was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSubmissionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          description: The client exceeded the submission rate limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: formora_session
...
  parameters:
...
    FormId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
  responses:
    ValidationError:
      description: The request did not match its boundary schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthenticationRequired:
      description: A valid session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PermissionDenied:
      description: The user does not have the required workspace permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
          minLength: 1
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
    Workspace:
      type: object
      required: [id, name, role, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, editor, viewer]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [workspace]
          properties:
            workspace:
              $ref: "#/components/schemas/Workspace"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    WorkspaceListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [workspaces]
          properties:
            workspaces:
              type: array
              items:
                $ref: "#/components/schemas/Workspace"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    FormField:
      type: object
      required: [id, type, label, required, placeholder, options]
      properties:
        id:
          type: string
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
        type:
          type: string
          enum: [text, textarea, email, number, select, checkbox]
        label:
          type: string
          minLength: 1
          maxLength: 120
        required:
          type: boolean
        placeholder:
          type: string
          maxLength: 200
        options:
          type: array
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 120
    FormDefinition:
      type: object
      required: [schemaVersion, fields]
      properties:
        schemaVersion:
          type: integer
          const: 1
        fields:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FormField"
...
    ManagedSubmissionSummary:
      type: object
      required: [id, publicationVersion, answeredFields, submittedAt]
      properties:
        id:
          type: string
        publicationVersion:
          type: integer
          minimum: 1
        answeredFields:
          type: integer
          minimum: 0
        submittedAt:
          type: string
          format: date-time
    ManagedSubmissionListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [submissions]
          properties:
            submissions:
              type: array
              items:
                $ref: "#/components/schemas/ManagedSubmissionSummary"
        meta:
          allOf:
            - $ref: "#/components/schemas/ResponseMeta"
            - type: object
              required: [page, pageSize, totalItems, totalPages, availableVersions]
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                availableVersions:
                  type: array
                  items:
                    type: integer
                    minimum: 1
...
    PublicForm:
      type: object
      required: [slug, title, requiresPassword, publicationVersion, definition]
      properties:
        slug:
          type: string
        title:
          type: string
        requiresPassword:
          type: boolean
        publicationVersion:
          type: integer
          minimum: 1
        definition:
          oneOf:
            - $ref: "#/components/schemas/FormDefinition"
            - type: "null"
    PublicFormResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [form]
          properties:
            form:
              $ref: "#/components/schemas/PublicForm"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    PublicFormAccessResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [form, accessToken, expiresAt]
          properties:
            form:
              $ref: "#/components/schemas/PublicForm"
            accessToken:
              type: string
            expiresAt:
              type: string
              format: date-time
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    PublicSubmissionRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          maxProperties: 100
          additionalProperties: true
    PublicSubmissionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [submissionId, submittedAt]
          properties:
            submissionId:
              type: string
            submittedAt:
              type: string
              format: date-time
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...